+++ backend/danswer/server/features/persona/api.py
@@ -1,4 +1,3 @@
-import json
 import uuid
 from uuid import UUID
 
@@ -14,12 +13,9 @@ from danswer.auth.users import current_admin_user
 from danswer.auth.users import current_curator_or_admin_user
 from danswer.auth.users import current_limited_user
 from danswer.auth.users import current_user
-from danswer.configs.chat_configs import NUM_PERSONA_PROMPT_GENERATION_CHUNKS
-from danswer.configs.chat_configs import NUM_PERSONA_PROMPTS
 from danswer.chat.prompt_builder.utils import build_dummy_prompt
 from danswer.configs.constants import FileOrigin
 from danswer.configs.constants import NotificationType
-from danswer.db.document_set import get_document_sets_by_ids
 from danswer.db.engine import get_session
 from danswer.db.models import StarterMessageModel as StarterMessage
 from danswer.db.models import User
@@ -37,16 +33,11 @@ from danswer.db.persona import update_persona_category
 from danswer.db.persona import update_persona_public_status
 from danswer.db.persona import update_persona_shared_users
 from danswer.db.persona import update_persona_visibility
-from danswer.document_index.document_index_utils import get_both_index_names
-from danswer.document_index.factory import get_default_document_index
 from danswer.file_store.file_store import get_default_file_store
 from danswer.file_store.models import ChatFileType
-from danswer.llm.factory import get_default_llms
-from danswer.prompts.starter_messages import PERSONA_STARTER_MESSAGE_CREATION_PROMPT
-from danswer.search.models import IndexFilters
-from danswer.search.models import InferenceChunk
-from danswer.search.postprocessing.postprocessing import cleanup_chunks
-from danswer.search.preprocessing.access_filters import build_access_filters_for_user
+from danswer.secondary_llm_flows.starter_message_creation import (
+    generate_starter_messages,
+)
 from danswer.server.features.persona.models import CreatePersonaRequest
 from danswer.server.features.persona.models import GenerateStarterMessageRequest
 from danswer.server.features.persona.models import ImageGenerationToolStatus
@@ -58,8 +49,6 @@ from danswer.server.features.persona.models import PromptTemplateResponse
 from danswer.server.models import DisplayPriorityRequest
 from danswer.tools.utils import is_image_generation_available
 from danswer.utils.logger import setup_logger
-from danswer.utils.threadpool_concurrency import FunctionCall
-from danswer.utils.threadpool_concurrency import run_functions_in_parallel
 
 
 logger = setup_logger()
@@ -381,79 +370,78 @@ def build_final_template_prompt(
     )
 
 
-def get_random_chunks_from_doc_sets(
-    doc_sets: list[str], db_session: Session, user: User | None = None
-) -> list[InferenceChunk]:
-    curr_ind_name, sec_ind_name = get_both_index_names(db_session)
-    document_index = get_default_document_index(curr_ind_name, sec_ind_name)
+# def generate_starter_messages(
+#     name: str,
+#     description: str,
+#     instructions: str,
+#     document_set_ids: list[int],
+#     db_session: Session,
+#     user: User | None,
+# ) -> list[StarterMessage]:
+#     start_message_generation_prompt = PERSONA_STARTER_MESSAGE_CREATION_PROMPT.format(
+#         name=name, description=description, instructions=instructions
+#     )
+#     _, fast_llm = get_default_llms(temperature=1.3)
 
-    acl_filters = build_access_filters_for_user(user, db_session)
-    filters = IndexFilters(document_set=doc_sets, access_control_list=acl_filters)
 
-    chunks = document_index.random_retrieval(
-        filters=filters, num_to_retrieve=NUM_PERSONA_PROMPT_GENERATION_CHUNKS
-    )
-    return cleanup_chunks(chunks)
+#     if document_set_ids:
+#         document_sets = get_document_sets_by_ids(
+#             document_set_ids=document_set_ids,
+#             db_session=db_session,
+#         )
+
+#         chunks = get_random_chunks_from_doc_sets(
+#             doc_sets=[doc_set.name for doc_set in document_sets],
+#             db_session=db_session,
+#             user=user,
+#         )
+
+#         # Add example content context to the prompt
+#         chunk_contents = "\n".join(chunk.content.strip() for chunk in chunks)
+
+#         start_message_generation_prompt += (
+#             "\n\nExample content this assistant has access to:\n"
+#             "'''\n"
+#             f"{chunk_contents}"
+#             "\n'''"
+#         )
 
 
-def generate_starter_messages(
-    name: str,
-    description: str,
-    instructions: str,
-    document_set_ids: list[int],
-    db_session: Session,
-    user: User | None,
-) -> list[StarterMessage]:
-    start_message_generation_prompt = PERSONA_STARTER_MESSAGE_CREATION_PROMPT.format(
-        name=name, description=description, instructions=instructions
-    )
-    _, fast_llm = get_default_llms(temperature=1.3)
+#     provider = fast_llm.config.model_provider
+#     model = fast_llm.config.model
 
-    if document_set_ids:
-        document_sets = get_document_sets_by_ids(
-            document_set_ids=document_set_ids,
-            db_session=db_session,
-        )
+#     #     params = get_supported_openai_params(model="anthropic.claude-3", custom_llm_provider="bedrock")
 
-        chunks = get_random_chunks_from_doc_sets(
-            doc_sets=[doc_set.name for doc_set in document_sets],
-            db_session=db_session,
-            user=user,
-        )
+#     #     assert "response_format" in params
+#     params = get_supported_openai_params(model=model, custom_llm_provider=provider)
+#     supports_structured_output = "response_format" in params
 
-        # Add example content context to the prompt
-        chunk_contents = "\n".join(chunk.content.strip() for chunk in chunks)
 
-        start_message_generation_prompt += (
-            "\n\nExample content this assistant has access to:\n"
-            "'''\n"
-            f"{chunk_contents}"
-            "\n'''"
-        )
+#     prompts: list[StarterMessage] = []
+#     functions = [
+#         FunctionCall(
+#             fast_llm.invoke,
+#             (start_message_generation_prompt, None, None, StarterMessage),
+#         )
+#         for _ in range(NUM_PERSONA_PROMPTS)
+#     ]
 
-    prompts: list[StarterMessage] = []
-    functions = [
-        FunctionCall(
-            fast_llm.invoke,
-            (start_message_generation_prompt, None, None, StarterMessage),
-        )
-        for _ in range(NUM_PERSONA_PROMPTS)
-    ]
+#     results = run_functions_in_parallel(function_calls=functions)
+#     for response in results.values():
+#         try:
+#             response_dict = json.loads(response.content)
+#             starter_message = StarterMessage(
+#                 name=response_dict["name"],
+#                 description=response_dict["description"],
+#                 message=response_dict["message"],
+#             )
+#             prompts.append(starter_message)
+#         except json.JSONDecodeError as e:
+#             logger.error(f"JSON decoding failed: {e}")
+#             continue
 
-    results = run_functions_in_parallel(function_calls=functions)
-    for response in results.values():
-        try:
-            response_dict = json.loads(response.content)
-            starter_message = StarterMessage(
-                name=response_dict["name"],
-                description=response_dict["description"],
-                message=response_dict["message"],
-            )
-            prompts.append(starter_message)
-        except json.JSONDecodeError as e:
-            logger.error(f"JSON decoding failed: {e}")
-            continue
-    return prompts
+
+#     return prompts
 
 
 @basic_router.post("/assistant-prompt-refresh")
